+++ core/server/middleware/index.js
@@ -1,37 +1,18 @@
 var debug           = require('debug')('ghost:middleware'),
-    path            = require('path'),
 
     // app requires
     config          = require('../config'),
-    helpers         = require('../helpers'),
     logging         = require('../logging'),
-    routes          = require('../routes'),
-    storage         = require('../storage'),
-    utils           = require('../utils'),
-
-    // This should probably be an internal app
-    sitemapHandler  = require('../data/xml/sitemap/handler'),
 
     // middleware
     compress        = require('compression'),
     netjet          = require('netjet'),
 
     // local middleware
-    cacheControl    = require('./cache-control'),
-    checkSSL        = require('./check-ssl'),
-    errorHandler    = require('./error-handler'),
-    ghostLocals     = require('./ghost-locals'),
-    maintenance     = require('./maintenance'),
-    prettyURLs      = require('./pretty-urls'),
-    serveSharedFile = require('./serve-shared-file'),
-    staticTheme     = require('./static-theme'),
-    themeHandler    = require('./theme-handler');
+    ghostLocals     = require('./ghost-locals');
 
 module.exports = function setupMiddleware(parentApp) {
     debug('Middleware start');
-
-    // @TODO find this a proper home
-    var blogApp = require('express')();
 
     // ## Global settings
 
@@ -78,78 +59,13 @@ module.exports = function setupMiddleware(parentApp) {
     // This sets global res.locals which are needed everywhere
     parentApp.use(ghostLocals);
 
-    // ## App - specific code
-    // set the view engine
-    blogApp.set('view engine', 'hbs');
-
+    // @TODO where should this live?!
     // Load helpers
-    helpers.loadCoreHelpers();
+    require('../helpers').loadCoreHelpers();
     debug('Helpers done');
 
-    // Theme middleware
-    // rightly or wrongly currently comes before theme static assets
-    // @TODO revisit where and when these are needed
-    blogApp.use(themeHandler.updateActiveTheme);
-    blogApp.use(themeHandler.configHbsForContext);
-    debug('Themes done');
-
-    // Static content/assets
-    // @TODO make sure all of these have a local 404 error handler
-    // Favicon
-    blogApp.use(serveSharedFile('favicon.ico', 'image/x-icon', utils.ONE_DAY_S));
-    // Ghost-Url
-    blogApp.use(serveSharedFile('shared/ghost-url.js', 'application/javascript', utils.ONE_HOUR_S));
-    blogApp.use(serveSharedFile('shared/ghost-url.min.js', 'application/javascript', utils.ONE_HOUR_S));
-    // Serve sitemap.xsl file
-    blogApp.use(serveSharedFile('sitemap.xsl', 'text/xsl', utils.ONE_DAY_S));
-    // Serve robots.txt if not found in theme
-    blogApp.use(serveSharedFile('robots.txt', 'text/plain', utils.ONE_HOUR_S));
-    // Serve blog images using the storage adapter
-    blogApp.use('/content/images', storage.getStorage().serve());
-
-    // Theme static assets/files
-    blogApp.use(staticTheme());
-    debug('Static content done');
-
-    // setup middleware for internal apps
-    // @TODO: refactor this to be a proper app middleware hook for internal & external apps
-    config.get('internalApps').forEach(function (appName) {
-        var app = require(path.join(config.get('paths').internalAppPath, appName));
-        if (app.hasOwnProperty('setupMiddleware')) {
-            app.setupMiddleware(blogApp);
-        }
-    });
-
-    // site map - this should probably be refactored to be an internal app
-    sitemapHandler(blogApp);
-    debug('Internal apps done');
-
-    // send 503 error page in case of maintenance
-    blogApp.use(maintenance);
-
-    // Force SSL if required
-    // must happen AFTER asset loading and BEFORE routing
-    blogApp.use(checkSSL);
-
-    // Add in all trailing slashes & remove uppercase
-    // must happen AFTER asset loading and BEFORE routing
-    blogApp.use(prettyURLs);
-
-    // ### Caching
-    // Blog frontend is cacheable
-    blogApp.use(cacheControl('public'));
-
-    debug('General middleware done');
-
-    // Set up Frontend routes (including private blogging routes)
-    blogApp.use(routes.frontend());
-
-    // ### Error handlers
-    blogApp.use(errorHandler.pageNotFound);
-    blogApp.use(errorHandler.handleHTMLResponse);
-
     // Mount the  apps on the parentApp
-    // Load the API
+    // API
     // @TODO: finish refactoring the API app
     // @TODO: decide what to do with these paths - config defaults? config overrides?
     parentApp.use('/ghost/api/v0.1/', require('../api/app')());
@@ -157,9 +73,8 @@ module.exports = function setupMiddleware(parentApp) {
     // ADMIN
     parentApp.use('/ghost', require('../admin')());
 
-    debug('Admin app & api done');
-
-    parentApp.use(blogApp);
+    // BLOG
+    parentApp.use(require('../blog')());
 
     debug('Middleware end');
 };
