--- core/server/routes/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-    frontend: require('./frontend')
-};
